+++ packages/dnr-rulesets/tasks/update-readme.ts
@@ -1,7 +1,7 @@
 import fs from 'fs';
 
-import { QUICK_FIXES_FILTER_ID, README_PATH } from './constants';
-import { getMetadata, type Metadata } from './metadata';
+import { QUICK_FIXES_FILTER_ID, README_PATH } from '../common/constants';
+import { downloadMetadata, type Metadata } from '../common/metadata';
 
 /**
  * Creates actual rulesets list and update existed README.md.
@@ -48,6 +48,6 @@ async function updateReadme(metadata: Metadata): Promise<void> {
     await fs.promises.writeFile(README_PATH, newReadme);
 }
 
-getMetadata()
+downloadMetadata()
     .then(updateReadme)
     .catch(console.error);