+++ packages/tsurlfilter/CHANGELOG.md
@@ -5,16 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [3.4.1] - 2025-07-02
 
 ### Added
 
 - Possibility to exclude unsafe rules from rulesets and save them to the metadata of rulesets
   for use "skip review" feature in CWS.
+- Export of `isSafeRule` from `declarative-converter-utils`.
 
 ### Fixed
 
-- CLI was not worked since to ESM builds.
+- CLI was not worked since switch to ESM builds.
+- `unsafeRulesCount` is added for rulesets after the conversion.
+
+[3.4.1]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tsurlfilter-v3.4.1
 
 ## [3.4.0] - 2025-06-06